Move score submission status enum to constants
