+++ app/constants/__init__.py
@@ -6,3 +6,4 @@ from . import gamemodes
 from . import mods
 from . import privileges
 from . import regexes
+from . import score_statuses
